ウサギのレアリティの調整を行った
ウサギとモチの出現の割合とウサギのレアリティの抽選確率の調整を行った。
それに伴い、抽選確率に関わる変数をint型からfloat形に変更した。

指定の回数だけウサギの抽選を連続して行うテストプログラムを作成した。
レアリティの抽選確率のパーセントをログに表示するようにした。

ウサギ自体の出現確率はまだ未調整。
他の人からの意見をもとに引き続き調整していく。

【確認ファイル】
ObjectLotteryMachine.cs
RabbitLotteryTest.cs
TowerObjectDataManager.cs
RabbitData.cs
RabbitLotteryMachine.cs
RabbitRarityData.cs

+++ Assets/Scripts/TowerDatas/Rabbit/RabbitData.cs
@@ -16,8 +16,8 @@ public sealed class RabbitData : BaseData
     // ウサギの出現率
     [SerializeField]
     [Range(0,100)]
-    int spawnRate = 0;
+    float spawnRate = 0;
 
     public int        Number { get { return number; } }
-    public int        SpawnRate    { get { return spawnRate;    } }
+    public float      SpawnRate    { get { return spawnRate;    } }
 }
